refactor: move db context stuff out from app
