+++ src/hooks/use-database.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
+import { Database } from '@nozbe/watermelondb'
+import { DatabaseContext } from '../lib/context'
 
-import { DatabaseContext } from '../App'
-
-export default () => useContext( DatabaseContext )
+export default (): Database => useContext( DatabaseContext )
