+++ src/lib/context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+import database from './database'
+
+export const DatabaseContext = createContext( database )
